fix: Optimizations

+++ src/client/java/org/infinite/libs/graphics/TexturedQuadRenderState.java
@@ -60,19 +60,19 @@ public record TexturedQuadRenderState(
 
   public void setupVertices(VertexConsumer vertices) {
     vertices
-        .vertex(this.pose(), (float) this.x1(), (float) this.y1())
+        .vertex(this.pose(), this.x1(), this.y1())
         .texture(this.u1(), this.v1())
         .color(this.color());
     vertices
-        .vertex(this.pose(), (float) this.x1(), (float) this.y2())
+        .vertex(this.pose(), this.x1(), this.y2())
         .texture(this.u1(), this.v2())
         .color(this.color());
     vertices
-        .vertex(this.pose(), (float) this.x2(), (float) this.y2())
+        .vertex(this.pose(), this.x2(), this.y2())
         .texture(this.u2(), this.v2())
         .color(this.color());
     vertices
-        .vertex(this.pose(), (float) this.x2(), (float) this.y1())
+        .vertex(this.pose(), this.x2(), this.y1())
         .texture(this.u2(), this.v1())
         .color(this.color());
   }
